Add files via upload
Considering the same revenue in 2015,
which is $5953.86 million, there is a 2.3% reduction in revenue for the newly established taxes to be paid. This accounts to forecast of a new revenue of $5816.92 million in 2015. Considering that all other costs of goods, research and development, operating income and others remain same in 2015, to maintain the same net earnings as the previous year 2014, a total of $136.94 million in net earnings is to be made up. To ensure that this drop in net earnings is made up, it is necessary to take action against certain sectors such as trying to increase the operational income, trying to cut down unnecessary other expenses, inventory holding costs, reducing costs of marketing by hiring only necessary sales people and using smart kiosks and online resources to market and sell our products, usage of a Kanban system at our warehouses and inventories, and using automation to reduce the need of workers in the long run, which in turn increases the rate at which we push the products to our customers and provide more sales and profit.